Please enter a commit message to explain why this merge is necessary,
 especially if it merges an updated upstream into a topic branch.

 Lines starting with '#' will be ignored, and an empty message aborts
 the commit.

+++ transport/tcp/endpoint.go
@@ -143,7 +143,7 @@ type endpoint struct {
 	addrsArr []*net.UDPAddr
 	addrsMap map[[6]byte]*net.UDPAddr
 	addrMu   sync.Mutex
-	addrIdx  uint8
+	addrIdx  int
 
 	subnetInited bool
 	subnetIP     global.Address
